+++ src/test/java/com/pragmaticcoders/checkout/domain/PromotionTest.java
@@ -2,6 +2,7 @@ package com.pragmaticcoders.checkout.domain;
 
 import org.junit.Test;
 
+import java.math.BigDecimal;
 import java.util.HashSet;
 import java.util.UUID;
 
@@ -18,13 +19,13 @@ public class PromotionTest {
             new HashSet<Item>() {{
                 add(mock(Item.class));
             }},
-            10
+            BigDecimal.valueOf(10)
         );
 
         promotion.getItems().add(mock(Item.class));
 
         assertEquals(uuid, promotion.getId());
         assertEquals(1, promotion.getItems().size());
-        assertEquals(Integer.valueOf(10), promotion.getDiscount());
+        assertEquals(BigDecimal.valueOf(10), promotion.getDiscount());
     }
 }
