+++ src/test/java/com/pragmaticcoders/checkout/view/ItemViewTest.java
@@ -2,6 +2,7 @@ package com.pragmaticcoders.checkout.view;
 
 import org.junit.Test;
 
+import java.math.BigDecimal;
 import java.util.ArrayList;
 import java.util.UUID;
 
@@ -17,13 +18,13 @@ public class ItemViewTest {
             uuid,
             "name",
             new ArrayList<ItemView.Price>() {{
-                add(new ItemView.Price(1, 20));
+                add(new ItemView.Price(1, BigDecimal.valueOf(20)));
             }}
         );
 
         assertEquals(uuid, view.getId());
         assertEquals("name", view.getName());
         assertEquals(Integer.valueOf(1), view.getPrices().get(0).getQuantity());
-        assertEquals(Integer.valueOf(20), view.getPrices().get(0).getPrice());
+        assertEquals(BigDecimal.valueOf(20), view.getPrices().get(0).getPrice());
     }
 }
