+++ src/test/java/com/pragmaticcoders/checkout/view/OrderViewTest.java
@@ -2,6 +2,7 @@ package com.pragmaticcoders.checkout.view;
 
 import org.junit.Test;
 
+import java.math.BigDecimal;
 import java.util.ArrayList;
 import java.util.UUID;
 
@@ -18,20 +19,20 @@ public class OrderViewTest {
         OrderView view = new OrderView(
             uuid,
             new ArrayList<OrderView.Item>() {{
-                add(new OrderView.Item(uuidPrice, 20, "foo", 10));
+                add(new OrderView.Item(uuidPrice, 20, "foo", BigDecimal.valueOf(10)));
             }},
-            200,
-            300,
+            BigDecimal.valueOf(200),
+            BigDecimal.valueOf(300),
             "test"
         );
 
         assertEquals(uuid, view.getId());
         assertEquals(uuidPrice, view.getItems().get(0).getId());
         assertEquals(Integer.valueOf(20), view.getItems().get(0).getQuantity());
-        assertEquals(Integer.valueOf(10), view.getItems().get(0).getPrice());
+        assertEquals(BigDecimal.valueOf(10), view.getItems().get(0).getPrice());
         assertEquals("foo", view.getItems().get(0).getName());
-        assertEquals(Integer.valueOf(300), view.getTotalPrice());
-        assertEquals(Integer.valueOf(200), view.getDiscount());
+        assertEquals(BigDecimal.valueOf(300), view.getTotalPrice());
+        assertEquals(BigDecimal.valueOf(200), view.getDiscount());
         assertEquals("test", view.getStatus());
 
     }
